requirements.txt
fastapi>=0.109
python-multipart>=0.0.6
uvicorn[standard]>=0.27
pydantic>=2
numpy>=1.24
pandas>=2
joblib>=1.3
scikit-learn>=1.3
lightgbm>=4
httpx>=0.26
pytest>=7.4
pdfplumber>=0.10
python-docx>=1.1
sqlalchemy>=2.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3
email-validator>=2.1
